refactor: Moved utils to separated files, code cleanup
