+++ src/shims-ext.d.ts
@@ -11,10 +11,4 @@ declare module 'vue/types/vue' {
     interface VueRoute {
         path: string;
     }
-
-    interface FormularioForm extends Vue {
-        name: string | boolean;
-        proxy: Record<string, any>;
-        hasValidationErrors(): Promise<boolean>;
-    }
 }
