+++ test/unit/utils/merge.test.js
